Fix CI file-length and positional-params violations.
Extract model display into agentui_model_display.go and use options structs for multi-param helpers.
